refactor: split TimelockTimeline into own file
