+++ src-gui/src/renderer/components/alert/SwapMightBeCancelledAlert.tsx
@@ -1,7 +1,7 @@
 import { makeStyles } from "@material-ui/core";
 import { Alert, AlertTitle } from "@material-ui/lab";
 import { useActiveSwapInfo } from "store/hooks";
-import { TimelockStatusAlert, TimelockTimeline } from "./SwapStatusAlert";
+import { TimelockStatusAlert, TimelockTimeline } from "./SwapStatusAlert/SwapStatusAlert";
 
 const useStyles = makeStyles((theme) => ({
   outer: {
